backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pandas
numpy
scikit-learn
python-dotenv
joblib
python-jose[cryptography]
passlib[bcrypt]